requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
streamlit==1.28.0
plotly==5.17.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
huggingface-hub==0.20.0